fin_news_digest/requirements.txt
apscheduler==3.10.4
feedparser==6.0.11
jinja2==3.1.4
requests==2.32.3
python-dotenv==1.0.1
